Feature/analytic ic (#5)
* First commit of analytic_IC branch

I added the dcmip f90 file that specifies the initial conditions for test cases 1, 2, and 3 and I modified the CMakeList.txt file in oops/src to include it in the build.  It successfully compiles but it have not been incorporated into the code yet.

* added qg_field_analytic_init

This is a new routine - it's not ready to be called yet but it exists!

* Added geometry key to qg_field_analytic_init()

* Added latitude, longitude, and area to qg_geom structure

...and initialized them in c_qg_geo_setup.  This follows the current version of the nicas_latest branch of oops, although I did not attempt an explicit merge.

* Added analytic_init member function to FieldsQG class

* Added analytic_init to config file

I made a new input file for testing called qg/test/testinput/analytic_init.json and I included a control structure in StateQG.cc to call the analytic_init() routine if this input field is present.

* Fleshed out analytic_init()

I completed a first go at the analytic_init() routine in qg_fields.F90, including an initialization of the fields from the dcmip routines.  I also incoporated the invent_state() routing into this analytic_init() framework. You can now call this routine by specifying State.analytic_init="baroclinic_instability" in the config file.  This version compiles but it is not yet tested.

* Cleaned up curl

Improved the calculation of the curl in the analytic_init() function in qg_fields.F90 - it's more readable now

* Added orography to qg_geom structure

* Defined test_qg_analytic_init

Currently this is just the same executable as test_qg_geometry but I'll build on this later.  This commit also includes the Periodic key word to the deriv() function in qg_fields.f90

* Included analytic_init.json in qg_test_input list

This version passes the TestGeometry.cc test but I have not yet checked the State.

* Now reading state variables from file

This was necessary because the dcmip routines need u and v.  Currently the state test is failing at calc_pv - still need to sort out the local boundary conditions

* This version passes test_qg_analytic_init

* Passes test_qg_truth and test_qg_state

Changed the calling sequence for calc_pv() to get it to work in the absence of latitudinal boundary conditions.  However, this seems to have broken some of the other tests.

* Reverted a backward-compatible format for the calc_pv argument list

...but this version still fails a number of tests, starting with test_qg_linearmodel

* Added lbc to all calc_pv() calls

* revert to old version of invent_state.f90

This made no difference - it is still failing the same tests

* reverted as much as possible to old version of qg_fields.F90

Now read_file() is unchanged.  The only change is the addition of analytic_init().  This version still passes test_qg_analytic_init and test_qg_state and test_qg_truth but it fails a number of other tests.

* revert to original calc_pv() c_qg_model.f90

This helped!  It's pasing more tests than before but still failing on some of the ufo things.  And, failing test_qg_analytic_init.

* reverted to original StateQG.cc

Still failing test_ufo_geovals, amsua, radiosonde, and seaice, along with of course test_qg_analytic_init

* Passes ufo and linear tests

This passes the tests that were failing before.  It turns out that failures were due at least in part to building with the --build=debug option.  This version restores some of my previous changes to StateQG.cc and qg_fields.f90

* Passes all tests

...with one exception, namely eckit_test_urlhandle.  However, the original develop version I branched from also failed this test so all is well with analytic_ic.

* Cleaned up a bit

* Added test_qg_aninit_fc

This test generates forecast data from the dcmip initial conditions that can be visualized with the python scripts

* Tweaked input

Altered layer depths to make it more consistent with the DCMIP IC.

* Added doxygen documentation to qg_fields.F90

* Added flat orography option

The DCMIP routines use flat orography so I made this an option in the "Model" section of the config file.

* Minor adjustments

* Eliminated CamelCase in qg_fields.F90

* Fixed more CamelCase